Updated folder structure
