+++ frontend/src/features/manageProfile/apiManager/manageProfileAPI.tsx
@@ -6,65 +6,8 @@ import {
   getCompanyIdFromSession,
 } from "../../../common/apiManager/httpRequestHandler";
 import { UserContextType } from "../../../common/authentication/types";
+import { CompanyProfile, CompanyAndUserRequest } from "../types/manageProfile";
 import { MANAGE_PROFILE_API, MANAGE_PROFILE_URL } from "./endpoints/endpoints";
-
-interface Address {
-  addressLine1: string;
-  addressLine2?: string;
-  city: string;
-  provinceCode: string;
-  countryCode: string;
-  postalCode: string;
-}
-
-interface Contact {
-  firstName: string;
-  lastName: string;
-  phone1: string;
-  phone1Extension?: string;
-  phone2?: string;
-  phone2Extension?: string;
-  email: string;
-  fax?: string;
-  city: string;
-  provinceCode: string;
-  countryCode: string;
-}
-
-export interface CompanyProfile {
-  companyId: string;
-  companyGUID: string;
-  clientNumber: string;
-  legalName: string;
-  companyAddress: Address;
-  mailingAddressSameAsCompanyAddress?: boolean;
-  mailingAddress?: Address;
-  email: string;
-  phone: string;
-  extension?: string;
-  fax?: string;
-  primaryContact: Contact;
-}
-
-export interface UserInformation extends Contact {
-  fax?: string;
-  userAuthGroup: string;
-}
-
-export interface CompanyAndUserRequest {
-  companyId: string;
-  companyGUID: string;
-  legalName: string;
-  companyAddress: Address;
-  mailingAddressSameAsCompanyAddress: boolean;
-  mailingAddress?: Address;
-  email: string;
-  phone: string;
-  extension?: string;
-  fax?: string;
-  primaryContact: Contact;
-  adminUser: UserInformation;
-}
 
 export const getCompanyInfo = async (): Promise<CompanyProfile> => {
   const url = new URL(MANAGE_PROFILE_API.COMPANIES);
