+++ frontend/src/features/permits/constants/constants.ts
